+++ mobile/src/components/Party/Nav.tsx
@@ -1,9 +1,10 @@
+import React, { FunctionComponent } from 'react';
 import { Avatar, Box, Grid } from '@material-ui/core';
 import Link from '@material-ui/core/Link';
 import Typography from '@material-ui/core/Typography';
 import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
 import { makeStyles } from '@material-ui/styles';
-import React from 'react';
+import Seats from '../Party/Seats';
 
 const useStyles = makeStyles({
     photo: {
@@ -13,56 +14,57 @@ const useStyles = makeStyles({
     }
 });
 
-export interface NavProps {
+export interface Props {
     name: string;
     logo: string;
     regionalSittingNum: number;
     nonRegionalSittingNum: number;
+    voteRate: string;
 }
 
-const Nav = ({
+const Nav: FunctionComponent<Props> = ({
     name,
     logo,
     regionalSittingNum,
-    nonRegionalSittingNum
-}: NavProps) => {
+    nonRegionalSittingNum,
+    voteRate
+}) => {
     const classes = useStyles();
     return (
-        <>
-            <Box m={1}>
-                <Box display="flex" alignItems="center" my={1}>
-                    <Link href={`/party`}>
-                        <KeyboardArrowLeft fontSize="large" />
-                    </Link>
-                    <Box>
-                        <Typography variant="h3" display="inline">
-                            {`${name} `}
-                        </Typography>
-                    </Box>
+        <Box m={1}>
+            <Box display="flex" alignItems="center" my={1}>
+                <Link href={`/parties`}>
+                    <KeyboardArrowLeft fontSize="large" />
+                </Link>
+                <Box>
+                    <Typography variant="h3" display="inline">
+                        {`${name} `}
+                    </Typography>
                 </Box>
-                <Box px={1} py={1}>
-                    <Grid container alignItems="center" spacing={2}>
-                        <Grid item>
-                            <Avatar className={classes.photo} src={logo} />
-                        </Grid>
-                        <Grid item>
-                            <Grid container direction="column">
-                                <Grid item>
-                                    <Typography variant="h5">
-                                        現任不分區立委 {regionalSittingNum} 席
-                                    </Typography>
-                                </Grid>
-                                <Grid item>
-                                    <Typography variant="h5">
-                                        區域立委 {nonRegionalSittingNum} 席
-                                    </Typography>
-                                </Grid>
+                <Seats num={regionalSittingNum} />
+            </Box>
+            <Box px={1} py={1}>
+                <Grid container alignItems="center" spacing={2}>
+                    <Grid item>
+                        <Avatar className={classes.photo} src={logo} />
+                    </Grid>
+                    <Grid item>
+                        <Grid container spacing={1} direction="column">
+                            <Grid item>
+                                <Typography variant="h5" color="textSecondary">
+                                    上屆政黨得票率 {voteRate}
+                                </Typography>
+                            </Grid>
+                            <Grid item>
+                                <Typography variant="h4" color="textSecondary">
+                                    現任不分區 {regionalSittingNum} 席 區域立委 {nonRegionalSittingNum} 席
+                                </Typography>
                             </Grid>
                         </Grid>
                     </Grid>
-                </Box>
+                </Grid>
             </Box>
-        </>
+        </Box>
     );
 };
